+++ src/briefy/common/db/model.py
@@ -48,6 +48,9 @@ class Security:
 class Base(Security):
     """Base Declarative model."""
 
+    __parent_attr__ = None
+    __additional_can_view_lr__ = []
+    __raw_acl__ = ()
     __session__ = None
     __exclude_attributes__ = [
         '_sa_instance_state', 'request', 'versions', 'path', 'can_view', 'type', 'local_roles'
@@ -298,7 +301,11 @@ class Base(Security):
         :param payload: Dictionary containing attributes and values
         :type payload: dict
         """
-        return cls(**payload)
+        obj = cls(**payload)
+        session = obj.__session__
+        session.add(obj)
+        session.flush()
+        return obj
 
 
 Base = declarative_base(cls=Base)