+++ src/briefy/common/db/models/local_role.py
@@ -13,9 +13,20 @@ class LocalRole(Timestamp, Identifiable, Base):
 
     __tablename__ = 'localroles'
 
+    item_type = sa.Column(
+        sa.String(255),
+        index=True,
+        nullable=False
+    )
+    """Item type.
+
+    Name of the entity -- as in its class name.
+    """
+
     item_id = sa.Column(
         UUID(as_uuid=True),
         sa.ForeignKey('items.id'),
+        index=True,
         nullable=False,
         info={
             'colanderalchemy': {
@@ -30,6 +41,7 @@ class LocalRole(Timestamp, Identifiable, Base):
     principal_id = sa.Column(
         UUID(as_uuid=True),
         nullable=False,
+        index=True,
         info={
             'colanderalchemy': {
                 'title': 'Principal UUID',
@@ -40,7 +52,7 @@ class LocalRole(Timestamp, Identifiable, Base):
     )
     """ID of the principal (user or group) which have the role in this context."""
 
-    role_name = sa.Column(sa.String(255), nullable=False)
+    role_name = sa.Column(sa.String(255), nullable=False, index=True)
     """Name of the role the permission granted."""
 
     def __repr__(self):