+++ tests/cache/test_cache_manager.py
@@ -1,20 +1,18 @@
 """Test CacheManager utility."""
+from briefy.common.db import Base
 from briefy.common.db.mixins import BaseMetadata
-from briefy.common.db.mixins import SubItemMixin
-from briefy.common.db.models import Item
+from briefy.common.db.mixins import Mixin
 from conftest import DBSession
 from dogpile.cache.backends.memory import MemoryBackend
 from dogpile.cache.backends.redis import RedisBackend
 
 import pytest
-import uuid
 
 
 dummy_cache_data = {
     'title': 'Dummy Cache Item',
     'updated_at': '2016-09-08T15:36:28.087123Z',
     'id': '6b6f0b2a-25ed-401c-8c65-3d4009e398ea',
-    'path': [uuid.UUID('6b6f0b2a-25ed-401c-8c65-3d4009e398ea')],
     'created_at': '2016-09-08T15:36:28.087112Z',
     'state': 'created',
 }
@@ -25,7 +23,7 @@ CACHE_BACKENDS = {
 }
 
 
-class DummyCache(BaseMetadata, SubItemMixin, Item):
+class DummyCache(BaseMetadata, Mixin, Base):
     """A content containing title, description and a slug."""
 
     __tablename__ = 'dummycaches'